pyproject.toml
[tool.poetry]
name = "magic-ball"
version = "0.1.0"
description = ""
authors = ["Dmitry Gorokhov <gorokhov86@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.1"
gunicorn = "^21.2.0"
make = "^0.1.6.post2"
python-dotenv = "^1.0.1"
flask-session = "^0.6.0"
flask-wtf = "^1.2.1"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
